+++ R/tm_g_km.R
@@ -334,7 +334,7 @@ tm_g_km <- function(label,
                     plot_width = NULL,
                     pre_output = NULL,
                     post_output = NULL) {
-
+  logger::log_info("Initializing tm_g_km")
   stop_if_not(
     is_character_single(label),
     is_character_single(dataname),