+++ R/tm_t_pp_prior_medication.R
@@ -168,6 +168,7 @@ tm_t_pp_prior_medication <- function(label,
                                      cmstdy = NULL,
                                      pre_output = NULL,
                                      post_output = NULL) {
+  logger::log_info("Initializing tm_t_pp_prior_medication")
   assert_that(is_character_single(label))
   assert_that(is_character_single(dataname))
   assert_that(is_character_single(parentname))